requirements.txt
# Core dependencies
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.26.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Hugging Face
huggingface-hub>=0.10.0

# Progress bars and visualization
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Streamlit demo
streamlit>=1.24.0

# FastAPI web demo
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
underthesea>=1.3.1
sentencepiece>=0.1.96
protobuf>=3.20.0
nltk>=3.8